+++ src/Providers/ConditionalProvidersServiceProvider.php
@@ -4,16 +4,7 @@ namespace SebastiaanLuca\ConditionalProviders\Providers;
 
 use Illuminate\Support\ServiceProvider;
 
-/**
- * Class EnvironmentProvidersServiceProvider
- *
- * Registers all service providers for the current environment as defined in the configuration.
- *
- * Inspired by https://mattstauffer.co/blog/conditionally-loading-service-providers-in-laravel-5 and others.
- *
- * @package SebastiaanLuca\ConditionalProviders\Providers
- */
-class EnvironmentProvidersServiceProvider extends ServiceProvider
+class ConditionalProvidersServiceProvider extends ServiceProvider
 {
     /**
      * Register the application services.
@@ -21,12 +12,13 @@ class EnvironmentProvidersServiceProvider extends ServiceProvider
     public function register()
     {
         $environment = $this->app->environment();
+
         $providers = $this->app['config']->get('app.' . $environment . '_providers');
-        
-        if (count($providers) === 0) {
+
+        if (! $providers) {
             return;
         }
-        
+
         foreach ($providers as $provider) {
             $this->app->register($provider);
         }