+++ tests/Helpers/MyLocalServiceProvider.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SebastiaanLuca\ConditionalProviders\Tests\Helpers;
+
+use Illuminate\Support\ServiceProvider;
+
+class MyLocalServiceProvider extends ServiceProvider
+{
+    /**
+     * Register the application services.
+     */
+    public function register()
+    {
+    }
+}